ヘッダメニューなどリファクタリング
