+++ assets/css/_base.scss
@@ -81,9 +81,3 @@ a:active {
     text-align: left;
   }
 }
-
-// wysiwyg-editor テーマ対応
-.g-theme-wysiwyg-editor {
-  --theme-wysiwyg-editor-toolbar-color: currentColor;
-  --theme-wysiwyg-editor-toolbar-active-color: currentColor;
-}
