+++ assets/css/_theme_default.scss
@@ -38,7 +38,7 @@
   .g-theme-contents-card {
     background-color: variables.$white;
   }
-  .g-theme-contets-item {
+  .g-theme-contents-item {
     &__header {
       background-color: #ccccff;
     }
